requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4

# Belge okuma
PyMuPDF==1.25.3
python-docx==1.1.2

# OCR (görsel belgeler için)
pytesseract==0.3.13
Pillow==11.1.0

# Embedding & Vektör Arama
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
torch==2.5.1
transformers==4.47.1

# Gemini
google-generativeai==0.8.4
